solid start support
