--- src/index.tsx
@@ -1,7 +1,0 @@
-/* @refresh reload */
-import { render } from "solid-js/web";
-
-import "./styles.css";
-import App from "./App";
-
-render(() => <App />, document.getElementById("root") as HTMLElement);
